chore(linux): port dsp and audiopassthru for linux compatibility Replace `long` with `int32_t` across all DSP and audiopassthru sources; `long` is 64-bit on linux x86_64 but 32-bit on windows, breaking fixed-layout DSP memory banks. Wrap windows-only APIs (COM init, dbghelp, NvOptimus exports) in `#ifdef _WIN32`; add linux crash handler via execinfo backtrace. Add portability shims in AudioPassthru.h (IMMDevice, WCHAR) for non-windows builds. Fix null-init pointer and swprintf bounds in DfxDspPreset.cpp; force headphone_on_ = IS_FALSE to prevent muffled output on speakers.

+++ dsp/DfxDspRegistry.cpp
@@ -27,7 +27,7 @@ along with this program.  If not, see <http://www.gnu.org/licenses/>.
 /*
 * Modeled after dfxg_WriteRegistrySessionLongValue() in dfxgSession.cpp.
 */
-int DfxDspPrivate::writeRegistrySessionLongValue(long l_value, wchar_t *wcp_key_name)
+int DfxDspPrivate::writeRegistrySessionLongValue(int32_t l_value, wchar_t *wcp_key_name)
 {
 	wchar_t wcp_key_value[DFXG_REGISTRY_BUFFER_LENGTH];
 	wchar_t wcp_full_key_path[PT_MAX_PATH_STRLEN];
